E_waste_classification/backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
ultralytics
groq
pinecone
sentence-transformers
pypdf
python-dotenv
tensorflow